pyproject.toml
[tool.poetry]
name = "jupyter2hashnode"
version = "0.1.6"
description = "Export from jupyter notebook into hashnode blog articles"
authors = ["Tiago Patricio Santos <tiagopatriciosantos@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/tiagopatriciosantos/jupyter2hashnode"

[tool.poetry.dependencies]
python = "^3.7"
nbconvert = {extras = ["all"], version = "^7.2.9"}
python-dotenv = "^0.21.1"
typer = {extras = ["all"], version = "^0.7.0"}
requests = "^2.28.2"
Pillow = "^9.4.0"

[tool.poetry.scripts]
rick-portal-gun = "jupyter2hashnode.cli:app"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

